Fixed main and container styles. Created new .js file and moved some files into folder
